requirements.txt
beautifulsoup4==4.8.0
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
djangorestframework==3.10.2
idna==2.8
lxml==4.4.0
pytz==2019.1
requests==2.22.0
soupsieve==1.9.2
sqlparse==0.3.0
urllib3==1.25.3
